only runs the job if the temp folder exists

+++ src/bootup/cron_jobs/delete_old_temp_folders.js
@@ -4,6 +4,7 @@ const Logger = require(Pathfinder.absPathInSrcFolder("utils/logger.js")).Logger;
 const tmpLocation = Pathfinder.absPathInApp("temp/");
 const exec = require("child_process").exec;
 const isNull = require(Pathfinder.absPathInSrcFolder("/utils/null.js")).isNull;
+const fs = require("fs");
 
 const deleteOldTempFolders = function (app, callback)
 {
@@ -14,31 +15,38 @@ const deleteOldTempFolders = function (app, callback)
         //deletes resources older than two hours inside the temp folder
         //find /Users/nelsonpereira/Desktop/infolaRepos/dendroRepo/dendro/temp/ -not -newermt '-7200 seconds' -mindepth 1 -delete
         //TODO look for a way to do this using nodejs code only -> so that it runs on machines other than macOs and Ubunto
-        exec("find " + tmpLocation + " -not -newermt '-7200 seconds' -mindepth 1", function(err, stdout, stderr)
-        {
-            if(!isNull(err))
+        if (fs.existsSync(tmpLocation)) {
+            exec("find " + tmpLocation + " -not -newermt '-7200 seconds' -mindepth 1", function(err, stdout, stderr)
             {
-                const errorMsg = "Error finding old resources in the delete_old_temp_folders job: " + err;
-                cb(true, errorMsg);
-            }
-            else
-            {
-                resourcesToDelete = stdout;
-                exec("find " + tmpLocation + " -not -newermt '-7200 seconds' -mindepth 1 -delete", function(err, stdout, stderr)
+                if(!isNull(err))
                 {
-                    if(!isNull(err))
+                    const errorMsg = "Error finding old resources in the delete_old_temp_folders job: " + err;
+                    cb(true, errorMsg);
+                }
+                else
+                {
+                    resourcesToDelete = stdout;
+                    exec("find " + tmpLocation + " -not -newermt '-7200 seconds' -mindepth 1 -delete", function(err, stdout, stderr)
                     {
-                        const errorMsg = "Error executing the delete_old_temp_folders job: " + err;
-                        cb(true, errorMsg);
-                    }
-                    else
-                    {
-                        const msg = "Executed the delete_old_temp_folders job successfully, deleted: " + resourcesToDelete;
-                        cb(null, msg);
-                    }
-                });
-            }
-        });
+                        if(!isNull(err))
+                        {
+                            const errorMsg = "Error executing the delete_old_temp_folders job: " + err;
+                            cb(true, errorMsg);
+                        }
+                        else
+                        {
+                            const msg = "Executed the delete_old_temp_folders job successfully, deleted: " + resourcesToDelete;
+                            cb(null, msg);
+                        }
+                    });
+                }
+            });
+        }
+        else
+        {
+            const errorMsg = "Error executing the delete_old_temp_folders job: " + "Temp folder location does not exist, did you accidentally deleted it?";
+            cb(true, errorMsg);
+        }
     };
     try {
         //Every hour
